# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Catálogo de Restaurantes",
        "description": "API completa para gestión de restaurantes, categorías de menú y productos",
        "contact": {
            "name": "Equipo de Desarrollo",
            "email": "admin@catalogo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local de Desarrollo"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener lista de todos los productos",
                "operationId": "496dc08306524d77f18d124d1d6838aa",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de productos obtenida exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "48b59f126c849beb4d4de4ff83b7945f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "restaurante_id",
                                    "categoria_id",
                                    "nombre",
                                    "precio",
                                    "disponible"
                                ],
                                "properties": {
                                    "restaurante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hamburguesa Especial"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Deliciosa hamburguesa con queso y vegetales"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12.5
                                    },
                                    "disponible": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "imagen": {
                                        "description": "Imagen del producto (JPEG, PNG, JPG, GIF, máximo 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Producto creado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Errores de validación"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "restaurante_id": [
                                                    "El ID del restaurante es obligatorio"
                                                ],
                                                "precio": [
                                                    "El precio debe ser mayor a 0"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
        "/api/productos/restaurante/{restauranteId}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener productos por restaurante",
                "operationId": "53518b05c0063762c8c22124417ad353",
                "parameters": [
                    {
                        "name": "restauranteId",
                        "in": "path",
                        "description": "ID del restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Productos del restaurante obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/categoria/{categoriaId}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener productos por categoría",
                "operationId": "3deb5883d4653ef90bfb48a5f3b818cc",
                "parameters": [
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Productos de la categoría obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {}
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Producto": {
                "title": "Producto",
                "description": "Modelo de Producto del catálogo",
                "required": [
                    "restaurante_id",
                    "categoria_id",
                    "nombre",
                    "precio",
                    "disponible"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del producto",
                        "type": "integer",
                        "format": "int64"
                    },
                    "restaurante_id": {
                        "description": "ID del restaurante al que pertenece el producto",
                        "type": "integer"
                    },
                    "categoria_id": {
                        "description": "ID de la categoría del producto",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "maxLength": 255
                    },
                    "descripcion": {
                        "description": "Descripción detallada del producto",
                        "type": "string",
                        "nullable": true
                    },
                    "precio": {
                        "description": "Precio del producto",
                        "type": "number",
                        "format": "float",
                        "minimum": 0
                    },
                    "disponible": {
                        "description": "Estado de disponibilidad del producto",
                        "type": "boolean"
                    },
                    "imagen": {
                        "description": "Ruta de la imagen del producto",
                        "type": "string",
                        "nullable": true
                    },
                    "imagen_url": {
                        "description": "URL completa de la imagen del producto",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}